+++ Assets/Animations/AdoptCatAnimController.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd95f233e2527f742aa210508ce04b08
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
