+++ Assets/Animations/Cat_Idle_Anim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 435e157a70433d6448ac93ecd1afd758
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
