+++ Assets/Animations/Cat_Jump_Anim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e7effe0315627db408aa8e4708ab3baf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
