+++ Assets/Animations/Cat_Stretch_Anim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e51613e2ed951b46ae97f6b09371e00
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
