+++ Assets/Animations/LivingRoomCatAnimController.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a769c8f438d6ad41a81c38c7a71d050
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
