+++ Assets/Animations/Walk_cycle_Anim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cae3a5f6d1e217b46b708820ea98897b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
