+++ Assets/Materials/Cat.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 409c984e3c449f4488fe8ee345203e88
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
