+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.3.3f1
+m_EditorVersion: 2018.3.4f1
